add to server, add stack overflow guard
