+++ packages/svelte/tests/snapshot/samples/static-template-expression-evaluation/_expected/server/index.svelte.js
@@ -7,7 +7,7 @@ export default function Static_template_expression_evaluation($$payload) {
 	let count = 0;
 
 	function Component() {} // placeholder component
-	$$payload.out += `<h1>Hello, ${$.escape(name)}!</h1> <p>${$.escape(a)} + ${$.escape(b)} = ${$.escape(a + b)}</p> <button>Count is ${$.escape(count)}</button> <p>1 + 2 = ${$.escape(1 + 2)}</p> <p>Sum is ${$.escape((a, b, a + b))}</p> <p>${$.escape(a === 1 ? a : b)}</p> `;
+	$$payload.out += `<h1>Hello, ${$.escape(name)}!</h1> <p>${$.escape(a)} + ${$.escape(b)} = ${$.escape(a + b)}</p> <button>Count is ${$.escape(count)}</button> <p>1 + 2 = 3</p> <p>Sum is ${$.escape((a, b, a + b))}</p> <p>${$.escape(a === 1 ? a : b)}</p> `;
 	Component($$payload, { a, count, c: a + b });
 	$$payload.out += `<!---->`;
 }
